cutting down
cut down on the amount of files by combining pdf parse and pdf sort
