+++ PDFParse/pdfparse.js
@@ -42,7 +42,7 @@ var pdf = require("pdf-parse");
 var readlineSync = require("readline-sync");
 function convertPdfToJson() {
     return __awaiter(this, void 0, void 0, function () {
-        var pdfPath, dataBuffer, data, textContent_1, keywords, jsonPath, error_1;
+        var pdfPath, dataBuffer, data, textContent_1, emailRegex, phoneRegex, addressRegex, emails, phoneNumbers, addresses, keywords, jsonPath, parsedInfo, outputJsonPath, error_1;
         return __generator(this, function (_a) {
             switch (_a.label) {
                 case 0:
@@ -58,16 +58,42 @@ function convertPdfToJson() {
                 case 1:
                     data = _a.sent();
                     textContent_1 = data.text;
-                    keywords = ['experience', 'skills', 'education', 'projects', 'achievements'];
+                    emailRegex = /\b[\w\.-]+@[\w\.-]+\.\w+\b/g;
+                    phoneRegex = /\b\d{3}[-.\s]?\d{3}[-.\s]?\d{4}\b/g;
+                    addressRegex = /\b\d+\s[\w\s]+,\s[\w\s]+,\s[\w\s\d]+\b/g;
+                    emails = textContent_1.match(emailRegex);
+                    if (emails) {
+                        textContent_1 = textContent_1.replace(emailRegex, '');
+                        textContent_1 += '\nEmail Addresses:\n' + emails.join('\n') + '\n';
+                    }
+                    phoneNumbers = textContent_1.match(phoneRegex);
+                    if (phoneNumbers) {
+                        textContent_1 = textContent_1.replace(phoneRegex, '');
+                        textContent_1 += '\nPhone Numbers:\n' + phoneNumbers.join('\n') + '\n';
+                    }
+                    addresses = textContent_1.match(addressRegex);
+                    if (addresses) {
+                        textContent_1 = textContent_1.replace(addressRegex, '');
+                        textContent_1 += '\nAddresses:\n' + addresses.join('\n') + '\n';
+                    }
+                    keywords = ['experience', 'skills', 'education', 'projects', 'achievements', 'interests'];
                     // Highlight keywords in uppercase
                     keywords.forEach(function (keyword) {
                         var regex = new RegExp("\\b".concat(keyword, "\\b"), 'gi'); // Using RegExp for whole word matching
                         textContent_1 = textContent_1.replace(regex, function (match) { return match.toUpperCase(); });
                     });
+                    // Handle multi-word categories
+                    textContent_1 = textContent_1.replace(/\b(Skills & Interests|Projects & Activities)\b/gi, function (match) { return match.toUpperCase(); });
                     jsonPath = path.join(__dirname, 'output.json');
                     // Write modified text content to a JSON file
                     fs.writeFileSync(jsonPath, textContent_1);
-                    console.log('Conversion successful! Text saved to output.json');
+                    console.log('Conversion successful! Text saved to output.json.');
+                    parsedInfo = parseResume(jsonPath);
+                    outputJsonPath = path.join(__dirname, 'parsed_output.json');
+                    fs.writeFileSync(outputJsonPath, JSON.stringify(parsedInfo, null, 2));
+                    console.log('Parsed Information:');
+                    console.log(parsedInfo);
+                    console.log("Parsed information has been written to ".concat(outputJsonPath));
                     return [3 /*break*/, 3];
                 case 2:
                     error_1 = _a.sent();
@@ -77,6 +103,46 @@ function convertPdfToJson() {
             }
         });
     });
+}
+function isName(line, isFirstNonEmptyLine) {
+    var trimmedLine = line.trim();
+    return isFirstNonEmptyLine && trimmedLine.length > 0 ? trimmedLine : null;
+}
+function parseResume(jsonPath) {
+    var lines = fs.readFileSync(jsonPath, 'utf-8').split('\n');
+    var parsedInfo = [];
+    var currentCategory = null;
+    var isFirstNonEmptyLine = true;
+    lines.forEach(function (line) {
+        // Skip empty lines
+        if (!line.trim()) {
+            return;
+        }
+        // Check if the line is the name
+        var name = isName(line, isFirstNonEmptyLine);
+        if (name) {
+            parsedInfo.push({ category: 'Name', content: [name] });
+            isFirstNonEmptyLine = false; // Set to false after finding the first non-empty line
+            return;
+        }
+        // Identify section headers and set the current category
+        var sectionMatch = line.match(/^[A-Z\s]+$/);
+        if (sectionMatch) {
+            currentCategory = sectionMatch[0].trim();
+            return;
+        }
+        // Add content to the current category
+        if (currentCategory !== null) {
+            var index = parsedInfo.findIndex(function (info) { return info.category === currentCategory; });
+            if (index === -1) {
+                parsedInfo.push({ category: currentCategory, content: [line.trim()] });
+            }
+            else {
+                parsedInfo[index].content.push(line.trim());
+            }
+        }
+    });
+    return parsedInfo;
 }
 // Call the function to start the conversion
 convertPdfToJson();
