+++ PDFParse/pdfparse.ts
@@ -3,6 +3,11 @@ import * as path from 'path';
 import * as pdf from 'pdf-parse';
 import * as readlineSync from 'readline-sync';
 
+interface ParsedInfo {
+  category: string;
+  content: string[];
+}
+
 async function convertPdfToJson(): Promise<void> {
   try {
     // Get PDF file path from user input
@@ -68,9 +73,68 @@ async function convertPdfToJson(): Promise<void> {
     fs.writeFileSync(jsonPath, textContent);
 
     console.log('Conversion successful! Text saved to output.json.');
+
+    // Example usage of parseResume function
+    const parsedInfo = parseResume(jsonPath);
+
+    // Write the parsed information to a JSON file
+    const outputJsonPath = path.join(__dirname, 'parsed_output.json');
+    fs.writeFileSync(outputJsonPath, JSON.stringify(parsedInfo, null, 2));
+
+    console.log('Parsed Information:');
+    console.log(parsedInfo);
+    console.log(`Parsed information has been written to ${outputJsonPath}`);
   } catch (error) {
     console.error('Error converting PDF to JSON:', error.message);
   }
+}
+
+function isName(line: string, isFirstNonEmptyLine: boolean): string | null {
+  const trimmedLine = line.trim();
+  return isFirstNonEmptyLine && trimmedLine.length > 0 ? trimmedLine : null;
+}
+
+function parseResume(jsonPath: string): ParsedInfo[] {
+  const lines = fs.readFileSync(jsonPath, 'utf-8').split('\n');
+
+  const parsedInfo: ParsedInfo[] = [];
+  let currentCategory: string | null = null;
+  let isFirstNonEmptyLine = true;
+
+  lines.forEach(line => {
+    // Skip empty lines
+    if (!line.trim()) {
+      return;
+    }
+
+    // Check if the line is the name
+    const name = isName(line, isFirstNonEmptyLine);
+    if (name) {
+      parsedInfo.push({ category: 'Name', content: [name] });
+      isFirstNonEmptyLine = false; // Set to false after finding the first non-empty line
+      return;
+    }
+
+    // Identify section headers and set the current category
+    const sectionMatch = line.match(/^[A-Z\s]+$/);
+    if (sectionMatch) {
+      currentCategory = sectionMatch[0].trim();
+      return;
+    }
+
+    // Add content to the current category
+    if (currentCategory !== null) {
+      const index = parsedInfo.findIndex(info => info.category === currentCategory);
+
+      if (index === -1) {
+        parsedInfo.push({ category: currentCategory, content: [line.trim()] });
+      } else {
+        parsedInfo[index].content.push(line.trim());
+      }
+    }
+  });
+
+  return parsedInfo;
 }
 
 // Call the function to start the conversion
